refactor(wp): move getRelatedResearchAndWritingForVariable to Post
