+++ baker/GrapherBaker.tsx
@@ -14,11 +14,7 @@ import {
     merge,
     partition,
 } from "@ourworldindata/utils"
-import {
-    getRelatedResearchAndWritingForVariable,
-    isWordpressAPIEnabled,
-    isWordpressDBEnabled,
-} from "../db/wpdb.js"
+import { isWordpressAPIEnabled, isWordpressDBEnabled } from "../db/wpdb.js"
 import fs from "fs-extra"
 import * as lodash from "lodash"
 import { bakeGraphersToPngs } from "./GrapherImageBaker.js"
@@ -34,6 +30,7 @@ import {
     getPostEnrichedBySlug,
     getPostRelatedCharts,
     getRelatedArticles,
+    getRelatedResearchAndWritingForVariable,
 } from "../db/model/Post.js"
 import {
     JsonError,
