+++ db/model/Post.ts
@@ -18,15 +18,16 @@ import {
     snapshotIsBlockGraphQlApi,
     PostReference,
     Tag,
+    DataPageRelatedResearch,
 } from "@ourworldindata/types"
-import { uniqBy } from "@ourworldindata/utils"
+import { uniqBy, sortBy, memoize, orderBy } from "@ourworldindata/utils"
 import { Knex } from "knex"
-import { memoize, orderBy } from "lodash"
 import { BAKED_BASE_URL } from "../../settings/clientSettings.js"
 import { BLOG_SLUG } from "../../settings/serverSettings.js"
 import { GdocPost } from "./Gdoc/GdocPost.js"
 import { SiteNavigationStatic } from "../../site/SiteNavigation.js"
 import { decodeHTML } from "entities"
+import { RelatedResearchQueryResult } from "../wpdb"
 
 export const postsTable = "posts"
 
@@ -376,4 +377,139 @@ export const getPostTags = async (
         .select("tags.id", "tags.name")
         .where({ post_id: postId })
         .join("tags", "tags.id", "=", "post_tags.tag_id")
+}
+
+export const getRelatedResearchAndWritingForVariable = async (
+    variableId: number
+): Promise<DataPageRelatedResearch[]> => {
+    const wp_posts: RelatedResearchQueryResult[] = await db.queryMysql(
+        `-- sql
+            -- What we want here is to get from the variable to the charts
+            -- to the posts and collect different pieces of information along the way
+            -- One important complication is that the slugs that are used in posts to
+            -- embed charts can either be the current slugs or old slugs that are redirected
+            -- now.
+            select
+                distinct
+                pl.target as linkTargetSlug,
+                pl.componentType as componentType,
+                coalesce(csr.slug, c.slug) as chartSlug,
+                p.title as title,
+                p.slug as postSlug,
+                coalesce(csr.chart_id, c.id) as chartId,
+                p.authors as authors,
+                p.featured_image as thumbnail,
+                coalesce(pv.views_365d, 0) as pageviews,
+                'wordpress' as post_source,
+                (select coalesce(JSON_ARRAYAGG(t.name), JSON_ARRAY())
+                    from post_tags pt
+                    join tags t on pt.tag_id = t.id
+                    where pt.post_id = p.id
+                ) as tags
+            from
+                posts_links pl
+            join posts p on
+                pl.sourceId = p.id
+            left join charts c on
+                pl.target = c.slug
+            left join chart_slug_redirects csr on
+                pl.target = csr.slug
+            left join chart_dimensions cd on
+                cd.chartId = coalesce(csr.chart_id, c.id)
+            left join analytics_pageviews pv on
+                pv.url = concat('https://ourworldindata.org/', p.slug )
+            left join posts_gdocs pg on
+            	pg.id = p.gdocSuccessorId
+            left join posts_gdocs pgs on
+                pgs.slug = p.slug
+            left join post_tags pt on
+                pt.post_id = p.id
+            where
+                -- we want only urls that point to grapher charts
+                pl.linkType = 'grapher'
+                -- componentType src is for those links that matched the anySrcregex (not anyHrefRegex or prominentLinkRegex)
+                -- this means that only the links that are of the iframe kind will be kept - normal a href style links will
+                -- be disregarded
+                and componentType = 'src'
+                and cd.variableId = ?
+                and cd.property in ('x', 'y') -- ignore cases where the indicator is size, color etc
+                and p.status = 'publish' -- only use published wp posts
+                and p.type != 'wp_block'
+                and coalesce(pg.published, 0) = 0 -- ignore posts if the wp post has a published gdoc successor. The
+                                                  -- coalesce makes sure that if there is no gdoc successor then
+                                                  -- the filter keeps the post
+                and coalesce(pgs.published, 0) = 0 -- ignore posts if there is a gdoc post with the same slug that is published
+                      -- this case happens for example for topic pages that are newly created (successorId is null)
+                      -- but that replace an old wordpress page
+
+            `,
+        [variableId]
+    )
+
+    const gdocs_posts: RelatedResearchQueryResult[] = await db.queryMysql(
+        `-- sql
+            select
+                distinct
+                pl.target as linkTargetSlug,
+                pl.componentType as componentType,
+                coalesce(csr.slug, c.slug) as chartSlug,
+                p.content ->> '$.title' as title,
+                p.slug as postSlug,
+                coalesce(csr.chart_id, c.id) as chartId,
+                p.content ->> '$.authors' as authors,
+                p.content ->> '$."featured-image"' as thumbnail,
+                coalesce(pv.views_365d, 0) as pageviews,
+                'gdocs' as post_source,
+                (select coalesce(JSON_ARRAYAGG(t.name), JSON_ARRAY())
+                    from posts_gdocs_x_tags pt
+                    join tags t on pt.tagId = t.id
+                    where pt.gdocId = p.id
+                ) as tags
+            from
+                posts_gdocs_links pl
+            join posts_gdocs p on
+                pl.sourceId = p.id
+            left join charts c on
+                pl.target = c.slug
+            left join chart_slug_redirects csr on
+                pl.target = csr.slug
+            join chart_dimensions cd on
+                cd.chartId = coalesce(csr.chart_id, c.id)
+            left join analytics_pageviews pv on
+                pv.url = concat('https://ourworldindata.org/', p.slug )
+            left join posts_gdocs_x_tags pt on
+                pt.gdocId = p.id
+            where
+                pl.linkType = 'grapher'
+                and componentType = 'chart' -- this filters out links in tags and keeps only embedded charts
+                and cd.variableId = ?
+                and cd.property in ('x', 'y') -- ignore cases where the indicator is size, color etc
+                and p.published = 1
+                and p.content ->> '$.type' != 'fragment'`,
+        [variableId]
+    )
+
+    const combined = [...wp_posts, ...gdocs_posts]
+
+    // we could do the sorting in the SQL query if we'd union the two queries
+    // but it seemed easier to understand if we do the sort here
+    const sorted = sortBy(combined, (post) => -post.pageviews)
+
+    const allSortedRelatedResearch = sorted.map((post) => {
+        const parsedAuthors = JSON.parse(post.authors)
+        const parsedTags = post.tags !== "" ? JSON.parse(post.tags) : []
+
+        return {
+            title: post.title,
+            url: `/${post.postSlug}`,
+            variantName: "",
+            authors: parsedAuthors,
+            imageUrl: post.thumbnail,
+            tags: parsedTags,
+        }
+    })
+    // the queries above use distinct but because of the information we pull in if the same piece of research
+    // uses different charts that all use a single indicator we would get duplicates for the post to link to so
+    // here we deduplicate by url. The first item is retained by uniqBy, latter ones are discarded.
+    return uniqBy(allSortedRelatedResearch, "url")
 }
